requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
PyJWT==2.8.0
setuptools==71.0.1
sqlparse==0.5.1
tzdata==2024.1
wheel==0.43.0
